i-oic-rbac-using-obo/mcp_server_code/requirements.txt
fastmcp>=2.12.4,<3.0.0
mcp>=1.15.0,<2.0.0
pyjwt>=2.9.0,<3.0.0
cryptography>=41.0.0,<43.0.0
anyio>=4.0.0,<5.0.0
httpx>=0.27.0,<0.28.0
starlette>=0.37.0,<0.38.0
python-dotenv>=1.0.0,<2.0.0
